moved TestGenerateDenyBlockedSecurityRuleName to securityrule_test.go
